refactored testing configs into one file
